+++ src/main/java/com/zewdiemarket/ws/service/representation/ReviewRepresentation.java
@@ -17,7 +17,7 @@ public class ReviewRepresentation extends AbstractRepresentation implements Comp
 	
 	private long id;
 	private String reviewDetail;
-	private float rating;
+	private double rating;
 	private String customerName;
 	
 	public ReviewRepresentation() {}
@@ -46,10 +46,10 @@ public class ReviewRepresentation extends AbstractRepresentation implements Comp
 		this.customerName = customerName;
 	}
 
-	public float getRating(){
+	public double getRating(){
 		return rating;
 	}
-	public void setRating(float rating) {
+	public void setRating(double rating) {
 		this.rating = rating;
 	}
